impl_wasm_map/list macro rewrite + clippy fixes + CI
impl_wasm_map is now aware of whether the keys/values are WASM-exposable
and copy or not. There is now a away to specify the backing map e.g.
BTreeMap for preserve-encodings=false i.e. Byron, CIP25

Use of this macro now replaces hard-generated WASM Map (and list) types
to reduce boilerplate + allow a single common place to fix linting
errors (one of the main reasons for this change to be done in this PR -
but it was planned at some point anyway)

Fixes #234

The rest of clippy linting errors were also fixed and a check in the CI
for fmt + linting.

+++ multi-era/rust/src/byron/delegation/serialization.rs
@@ -30,8 +30,8 @@ impl Deserialize for ByronDelegation {
         read_len.read_elems(4)?;
         read_len.finish()?;
         (|| -> Result<_, DeserializeError> {
-            let epoch = Ok(raw.unsigned_integer()? as u64)
-                .map_err(|e: DeserializeError| e.annotate("epoch"))?;
+            let epoch =
+                Ok(raw.unsigned_integer()?).map_err(|e: DeserializeError| e.annotate("epoch"))?;
             let issuer =
                 Ok(raw.bytes()? as Vec<u8>).map_err(|e: DeserializeError| e.annotate("issuer"))?;
             let delegate = Ok(raw.bytes()? as Vec<u8>)
@@ -114,9 +114,9 @@ impl Deserialize for EpochRange {
         read_len.read_elems(2)?;
         read_len.finish()?;
         (|| -> Result<_, DeserializeError> {
-            let epoch_id = Ok(raw.unsigned_integer()? as u64)
+            let epoch_id = Ok(raw.unsigned_integer()?)
                 .map_err(|e: DeserializeError| e.annotate("epoch_id"))?;
-            let epoch_id2 = Ok(raw.unsigned_integer()? as u64)
+            let epoch_id2 = Ok(raw.unsigned_integer()?)
                 .map_err(|e: DeserializeError| e.annotate("epoch_id2"))?;
             match len {
                 cbor_event::Len::Len(_) => (),
